requirements.txt
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-phonenumber-field==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.10.3
idna==2.8
mongoengine==0.18.2
oauthlib==3.1.0
phonenumbers==8.10.21
Pillow==6.2.1
pymongo==3.9.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
